+++ test/cli/shared/watch.dart
@@ -64,6 +64,48 @@ void sharedTests(Future<TestProcess> runSass(Iterable<String> arguments)) {
 
           await d.file("out.css", "x {y: z}").validate();
         });
+
+        test("continues compiling after an error", () async {
+          await d.file("test1.scss", "a {b: }").create();
+          await d.file("test2.scss", "x {y: z}").create();
+
+          var sass =
+              await watch(["test1.scss:out1.css", "test2.scss:out2.css"]);
+          await expectLater(sass.stderr, emits('Error: Expected expression.'));
+          await expectLater(
+              sass.stderr, emitsThrough(contains('test1.scss 1:7')));
+          await expectLater(
+              sass.stdout, emitsThrough('Compiled test2.scss to out2.css.'));
+          await expectLater(sass.stdout, _watchingForChanges);
+          await sass.kill();
+
+          await d.nothing("out1.css").validate();
+          await d
+              .file("out2.css", equalsIgnoringWhitespace("x { y: z; }"))
+              .validate();
+        });
+
+        test("stops compiling after an error with --stop-on-error", () async {
+          await d.file("test1.scss", "a {b: }").create();
+          await d.file("test2.scss", "x {y: z}").create();
+
+          var sass = await watch([
+            "--stop-on-error",
+            "test1.scss:out1.css",
+            "test2.scss:out2.css"
+          ]);
+          await expectLater(
+              sass.stderr,
+              emitsInOrder([
+                'Error: Expected expression.',
+                emitsThrough(contains('test1.scss 1:7')),
+                emitsDone
+              ]));
+          await sass.shouldExit(65);
+
+          await d.nothing("out1.css").validate();
+          await d.nothing("out2.css").validate();
+        });
       });
 
       group("recompiles a watched file", () {
@@ -164,6 +206,28 @@ void sharedTests(Future<TestProcess> runSass(Iterable<String> arguments)) {
           await expectLater(
               sass.stderr, emitsThrough(contains('test.scss 1:7')));
           await sass.kill();
+
+          await d.nothing("out.css").validate();
+        });
+
+        test("stops compiling after an error with --stop-on-error", () async {
+          await d.file("test.scss", "a {b: c}").create();
+
+          var sass = await watch(["--stop-on-error", "test.scss:out.css"]);
+          await expectLater(
+              sass.stdout, emits('Compiled test.scss to out.css.'));
+          await expectLater(sass.stdout, _watchingForChanges);
+          await tickIfPoll();
+
+          await d.file("test.scss", "a {b: }").create();
+          await expectLater(
+              sass.stderr,
+              emitsInOrder([
+                'Error: Expected expression.',
+                emitsThrough(contains('test.scss 1:7')),
+                emitsDone
+              ]));
+          await sass.shouldExit(65);
 
           await d.nothing("out.css").validate();
         });